refactor(pubsub): Handling of XDP in ethernet.c plugin file
 - Added XDP functionalities in ethernet.c
 - New connection config param(enableXdpSocket) to handle XDP
