+++ include/open62541/config.h.in
@@ -29,7 +29,6 @@
 #cmakedefine UA_ENABLE_PUBSUB
 #cmakedefine UA_ENABLE_PUBSUB_FILE_CONFIG
 #cmakedefine UA_ENABLE_PUBSUB_ETH_UADP
-#cmakedefine UA_ENABLE_PUBSUB_ETH_UADP_XDP
 #cmakedefine UA_ENABLE_PUBSUB_DELTAFRAMES
 #cmakedefine UA_ENABLE_PUBSUB_INFORMATIONMODEL
 #cmakedefine UA_ENABLE_PUBSUB_INFORMATIONMODEL_METHODS
