--- plugins/include/open62541/plugin/pubsub_ethernet_xdp.h
@@ -1,19 +1,0 @@
-/* This work is licensed under a Creative Commons CCZero 1.0 Universal License.
- * See http://creativecommons.org/publicdomain/zero/1.0/ for more information.
- *
- *   Copyright (c) 2019-2020 Kalycito Infotech Private Limited
- */
-
-#ifndef UA_NETWORK_PUBSUB_XDP_ETHERNET_H_
-#define UA_NETWORK_PUBSUB_XDP_ETHERNET_H_
-
-#include <open62541/plugin/pubsub.h>
-
-_UA_BEGIN_DECLS
-
-UA_PubSubTransportLayer UA_EXPORT
-UA_PubSubTransportLayerEthernetXDP(void);
-
-_UA_END_DECLS
-
-#endif /* UA_NETWORK_PUBSUB_XDP_ETHERNET_H_ */
