+++ plugins/ua_pubsub_ethernet.c
@@ -23,11 +23,95 @@
 #include <linux/if_packet.h>
 #include <netinet/ether.h>
 #include <linux/version.h>
+
+/* RLIMIT */
+#include <sys/time.h>
+#include <sys/resource.h>
+
+/* XSK kernel headers */
+#include <linux/bpf.h>
+#include <linux/if_link.h>
+#include <linux/if_xdp.h>
+
+#if __has_include(<bpf/bpf.h>) && __has_include(<bpf/libbpf.h>) && __has_include(<bpf/xsk.h>)
+#define LIBBPF_EBPF
+/* Libbpf headers */
+#include <bpf/bpf.h>
+#include <bpf/libbpf.h>
+#ifndef asm
+#define asm __asm__
 #endif
+#include <bpf/xsk.h>
+#endif
+#endif
+
 #include "time.h"
+
+#ifndef ETHERTYPE_UADP
 #define ETHERTYPE_UADP                       0xb62c
+#endif
+
 #define MIN_ETHERNET_PACKET_SIZE_WITHOUT_FCS 60
 #define VLAN_HEADER_SIZE                     4
+
+#define VLAN_SHIFT                           13
+#define XDP_FRAME_SHIFT                      11 //match the frame size 2048
+
+// Receive buffer batch size
+#define BATCH_SIZE                           8
+
+#ifndef XDP_COPY
+#define XDP_COPY                             (1 << 1)
+#endif
+
+#ifndef XDP_FLAGS_SKB_MODE
+#define XDP_FLAGS_SKB_MODE                   (1U << 1)
+#endif
+
+#if defined(LIBBPF_EBPF)
+/* Theses structures shall be removed in the future XDP versions
+ * (when RT Linux and XDP are mainlined and stable) */
+typedef struct {
+    char                    *frames;
+    struct xsk_ring_prod    fq;
+    struct xsk_ring_cons    cq;
+    struct xsk_umem *       umem;
+    void                    *buffer;
+    UA_Int32                fd;
+} xdp_umem;
+
+typedef struct {
+    struct xsk_ring_cons rx_ring;
+    struct xsk_ring_prod tx_ring;
+
+    struct xsk_socket *xskfd;
+    UA_UInt32 bpf_prog_id;
+
+    UA_UInt32 cur_tx;
+    UA_UInt32 cur_rx;
+
+    xdp_umem  *umem;
+    UA_UInt32 outstanding_tx;
+    UA_UInt64 rx_npkts;
+    UA_UInt64 tx_npkts;
+    UA_UInt64 prev_rx_npkts;
+    UA_UInt64 prev_tx_npkts;
+    UA_UInt32 idx_rx;
+    UA_UInt32 idx_fq;
+} xdpsock;
+#endif
+
+/* Configuration parameters for xdp socket */
+typedef struct {
+    UA_UInt32 no_of_frames;
+    UA_UInt32 frame_size;
+    UA_UInt32 no_of_desc;
+    UA_UInt32 fill_queue_no_of_desc;
+    UA_UInt32 com_queue_no_of_desc;
+} xskconfparam;
+
+/* Default values for xdp socket parameters */
+xskconfparam xsk_default_values = {2048, 2048, 2048, 2048, 2048};
 
 /* Ethernet network layer specific internal data */
 typedef struct {
@@ -36,21 +120,23 @@ typedef struct {
     UA_Byte prio;
     UA_Byte ifAddress[ETH_ALEN];
     UA_Byte targetAddress[ETH_ALEN];
+    /* XDP related param - Linux specific */
+    UA_Boolean enableXdpSocket;
+#if defined(LIBBPF_EBPF)
+    UA_UInt32 xdp_flags;
+    xdpsock* xdpsocket;
+#endif
     /* Send the packets with txtime ability - Linux specific */
     UA_Boolean useSoTxTime;
 } UA_PubSubChannelDataEthernet;
 
-/* Structure for Logical link control based on 802.2 */
-typedef struct  {
-    UA_Byte dsap;   /* Destination Service Access Point */
-    UA_Byte ssap;   /* Source Service Access point */
-    UA_Byte ctrl_1; /* Control Field */
-    UA_Byte ctrl_2; /* Control Field */
-} llc_pdu;
-
 /* Additional socket options */
 typedef struct {
     UA_UInt32  *socketPriority;
+    /* XDP related param - Linux specific */
+    UA_UInt32 xdp_flags;
+    UA_UInt32 hw_receive_queue;
+    UA_UInt16 xdp_bind_flags;
     /* ETF related socket options - Linux specific */
     UA_Boolean enableSocketTxTime;
     UA_Int32   sotxtimeDeadlinemode;
@@ -111,6 +197,305 @@ UA_parseHardwareAddress(UA_String* target, UA_Byte* destinationMac) {
 
     return UA_STATUSCODE_GOOD;
 }
+
+#if defined(LIBBPF_EBPF)
+/**
+ * UMEM is associated to a netdev and a specific queue id of that netdev.
+ * It is created and configured (chunk size, headroom, start address and size) by using the XDP_UMEM_REG setsockopt system call.
+ * UMEM uses two rings: FILL and COMPLETION. Each socket associated with the UMEM must have an RX queue, TX queue or both.
+ */
+static xdp_umem *xdp_umem_configure(xskconfparam *xskparam) {
+    struct xsk_umem_config uconfig = {
+        .fill_size = xskparam->fill_queue_no_of_desc,
+        .comp_size = xskparam->com_queue_no_of_desc,
+        .frame_size = xskparam->frame_size,
+        .frame_headroom = XSK_UMEM__DEFAULT_FRAME_HEADROOM,
+    };
+    xdp_umem *umem;
+    int ret;
+    size_t sret;
+    UA_UInt32 frames_per_ring;
+    UA_UInt32 idx;
+
+    umem = (xdp_umem *)UA_calloc(1, sizeof(*umem));
+    if (!umem) {
+        return NULL;
+    }
+
+    if (posix_memalign(&umem->buffer, (size_t)getpagesize(),
+                       xskparam->no_of_frames * xskparam->frame_size) != 0) {
+        UA_LOG_ERROR(UA_Log_Stdout, UA_LOGCATEGORY_SERVER,
+                     "buffer allocation of UMEM failed");
+        return NULL;
+    }
+
+    ret = xsk_umem__create(&umem->umem, umem->buffer,
+                           xskparam->no_of_frames * xskparam->frame_size,
+                           &umem->fq, &umem->cq,
+                           &uconfig);
+    if (ret) {
+        UA_LOG_ERROR(UA_Log_Stdout, UA_LOGCATEGORY_SERVER,
+                     "PubSub XSK UMEM creation failed. Out of memory.");
+        return NULL;
+    }
+
+    /* Populate rx fill ring with addresses */
+    frames_per_ring = xskparam->fill_queue_no_of_desc;
+
+    sret = xsk_ring_prod__reserve(&umem->fq, frames_per_ring, &idx);
+    if (sret != frames_per_ring)
+        return NULL;
+
+    for (UA_UInt64 i = 0; i < frames_per_ring; i++)
+        *xsk_ring_prod__fill_addr(&umem->fq, idx++) = i * frames_per_ring;
+
+    xsk_ring_prod__submit(&umem->fq, frames_per_ring);
+
+    return umem;
+}
+
+
+/**
+ *  Configure AF_XDP socket to redirect frames to a memory buffer in a user-space application
+ *  XSK has two rings: the RX ring and the TX ring
+ *  A socket can receive packets on the RX ring and it can send packets on the TX ring
+ */
+static xdpsock *xsk_configure(xdp_umem *umem, UA_UInt32 hw_receive_queue,
+                              int ifindex, char *ifname,
+                              UA_UInt32 xdp_flags, UA_UInt16 xdp_bind_flags) {
+    xdpsock *xdp_socket;
+    int ret;
+    struct xsk_socket_config cfg;
+
+    xskconfparam *xskparam = (xskconfparam *)UA_calloc(1, (sizeof(xskconfparam)));
+    memcpy(xskparam, &xsk_default_values, (sizeof(xskconfparam)));
+
+    xdp_socket = (xdpsock *)UA_calloc(1, sizeof(*xdp_socket));
+    if (!xdp_socket) {
+        return NULL;
+    }
+
+    if (umem)
+        xdp_socket->umem = umem;
+    else
+        xdp_socket->umem = xdp_umem_configure(xskparam);
+
+    if (!xdp_socket->umem) {
+        UA_close(xsk_socket__fd(xdp_socket->xskfd));
+        bpf_set_link_xdp_fd(ifindex, -1, xdp_flags);
+        return NULL;
+    }
+
+    xdp_socket->outstanding_tx = 0;
+    xdp_socket->idx_rx = 0;
+    xdp_socket->idx_fq = 0;
+
+    cfg.rx_size = xskparam->fill_queue_no_of_desc;
+    cfg.tx_size = xskparam->com_queue_no_of_desc;
+
+    cfg.libbpf_flags = 0;
+    cfg.xdp_flags = xdp_flags;
+    cfg.bind_flags = xdp_bind_flags;
+
+    ret = xsk_socket__create(&xdp_socket->xskfd,
+                             ifname,
+                             hw_receive_queue,
+                             xdp_socket->umem->umem,
+                             &xdp_socket->rx_ring,
+                             &xdp_socket->tx_ring,
+                             &cfg);
+
+    if (ret == ENOTSUP) {
+        UA_LOG_ERROR(UA_Log_Stdout, UA_LOGCATEGORY_SERVER,
+                      "PubSub connection creation failed."
+                      " xsk_socket__create not supported.");
+        return NULL;
+    } else if (ret < 0) {
+        UA_LOG_ERROR(UA_Log_Stdout, UA_LOGCATEGORY_SERVER,
+                    "PubSub connection creation failed."
+                    " xsk_socket__create failed: %s", strerror(errno));
+        UA_close(xsk_socket__fd(xdp_socket->xskfd));
+        bpf_set_link_xdp_fd(ifindex, -1, xdp_flags);
+        return NULL;
+    }
+
+    ret = bpf_get_link_xdp_id(ifindex, &xdp_socket->bpf_prog_id, xdp_flags);
+    if (ret) {
+        UA_LOG_ERROR (UA_Log_Stdout, UA_LOGCATEGORY_SERVER,
+            "PubSub Connection creation failed. Unable to retrieve XDP program.");
+        UA_close(xsk_socket__fd(xdp_socket->xskfd));
+        bpf_set_link_xdp_fd(ifindex, -1, xdp_flags);
+        return NULL;
+    }
+
+    UA_free(xskparam);
+    return xdp_socket;
+}
+
+static UA_StatusCode
+UA_PubSubChannelEthernetXDP_open(UA_PubSubChannelDataEthernet *channelDataEthernet,
+                                 ethernetSocketOptions *sockOptions,
+                                 UA_NetworkAddressUrlDataType *address) {
+    struct rlimit resourcelimit = {RLIM_INFINITY, RLIM_INFINITY};
+    if (setrlimit(RLIMIT_MEMLOCK, &resourcelimit)) {
+        UA_LOG_ERROR(UA_Log_Stdout, UA_LOGCATEGORY_SERVER, "Set limit on the consumption of resources failed");
+        return UA_STATUSCODE_BADINTERNALERROR;
+    }
+
+    /* Create the UMEM and XDP sockets */
+    channelDataEthernet->xdpsocket = xsk_configure(NULL,
+                                                   sockOptions->hw_receive_queue,
+                                                   channelDataEthernet->ifindex,
+                                                   (char *)address->networkInterface.data,
+                                                   sockOptions->xdp_flags,
+                                                   sockOptions->xdp_bind_flags);
+    if(!channelDataEthernet->xdpsocket) {
+        UA_LOG_ERROR(UA_Log_Stdout, UA_LOGCATEGORY_SERVER, "XDP Socket configuration failed");
+        bpf_set_link_xdp_fd(channelDataEthernet->ifindex, -1,
+                            sockOptions->xdp_flags);
+        return UA_STATUSCODE_BADINTERNALERROR;
+    }
+
+    return UA_STATUSCODE_GOOD;
+}
+
+static UA_StatusCode
+UA_PubSubChannelEthernetXDP_send(UA_PubSubChannelDataEthernet *channelDataEthernet,
+                                 const UA_ByteString *buf) {
+    struct ether_header* ethHdr;
+    UA_Byte *bufSend, *ptr;
+    xdpsock *xdp_socket;
+    UA_UInt16 vlanTag;
+    UA_UInt32 idx;
+    size_t lenBuf;
+    ssize_t rc;
+
+    xdp_socket = channelDataEthernet->xdpsocket;
+
+    bufSend = (UA_Byte *) xsk_umem__get_data(xdp_socket->umem->buffer,
+                                             xdp_socket->cur_tx << XDP_FRAME_SHIFT);
+    ethHdr = (struct ether_header *) bufSend;
+
+    /* Set (own) source MAC address */
+    memcpy(ethHdr->ether_shost, channelDataEthernet->ifAddress, ETH_ALEN);
+    /* Set destination MAC address */
+    memcpy(ethHdr->ether_dhost, channelDataEthernet->targetAddress, ETH_ALEN);
+    /* Set ethertype */
+    lenBuf = sizeof(*ethHdr) + buf->length;
+    ptr = sizeof(*ethHdr) + bufSend;
+
+    if(channelDataEthernet->vid > 0) {
+        lenBuf += 4;
+        ethHdr->ether_type = htons(ETHERTYPE_VLAN);
+        vlanTag = (UA_UInt16) (channelDataEthernet->vid
+                  + (channelDataEthernet->prio << VLAN_SHIFT));
+
+        *((UA_UInt16 *) ptr) = htons(vlanTag);
+        ptr += sizeof(UA_UInt16);
+
+        *((UA_UInt16 *) ptr) = htons(ETHERTYPE_UADP);
+        ptr += sizeof(UA_UInt16);
+    } else {
+        ethHdr->ether_type = htons(ETHERTYPE_UADP);
+    }
+
+    if (lenBuf > xsk_default_values.frame_size)
+        return UA_STATUSCODE_BADINTERNALERROR;
+
+    /* copy payload of ethernet message */
+    memcpy(ptr, buf->data, buf->length);
+
+    if (xsk_ring_prod__reserve(&xdp_socket->tx_ring, 1, &idx) != 1){
+        UA_LOG_ERROR(UA_Log_Stdout, UA_LOGCATEGORY_SERVER,
+                     "PubSub connection send failed. xsk_prod_reserve failed.");
+        return UA_STATUSCODE_BADINTERNALERROR;
+    }
+
+    xsk_ring_prod__tx_desc(&xdp_socket->tx_ring, idx)->addr = xdp_socket->cur_tx << XDP_FRAME_SHIFT;
+    xsk_ring_prod__tx_desc(&xdp_socket->tx_ring, idx)->len = (UA_UInt32) lenBuf;
+
+    xsk_ring_prod__submit(&xdp_socket->tx_ring, 1);
+    xdp_socket->outstanding_tx += 1;
+
+    /* Increase the cur_tx pointer, rollover if exceed */
+    xdp_socket->cur_tx += 1;
+    xdp_socket->cur_tx %= xsk_default_values.frame_size;
+
+    rc = UA_sendto(xsk_socket__fd(xdp_socket->xskfd), NULL, 0, MSG_DONTWAIT, NULL, 0);
+    if (rc >= 0 || errno == ENOBUFS || errno == EAGAIN || errno == EBUSY) {
+        UA_UInt64 rcvd = xsk_ring_cons__peek(&xdp_socket->umem->cq, 1, &idx);
+        if (rcvd > 0) {
+            xsk_ring_cons__release(&xdp_socket->umem->cq, rcvd);
+            xdp_socket->outstanding_tx -= (UA_UInt32) rcvd;
+            xdp_socket->tx_npkts += (UA_UInt32) rcvd;
+        }
+
+        return UA_STATUSCODE_GOOD;
+    }
+
+    UA_LOG_ERROR(UA_Log_Stdout, UA_LOGCATEGORY_SERVER,
+                 "PubSub connection send failed."
+                 " XSK Send message failed: %s", strerror(errno));
+    return UA_STATUSCODE_BADINTERNALERROR;
+
+}
+
+static UA_StatusCode
+UA_PubSubChannelEthernetXDP_receive(UA_PubSubChannelDataEthernet *channelDataEthernet,
+                                    UA_ByteString *message) {
+    struct ether_header *eth_hdr;
+    xdpsock *xdp_socket;
+    UA_UInt64 addr;
+    UA_UInt64 ret;
+    UA_Byte *pkt;
+    ssize_t len;
+
+    xdp_socket = channelDataEthernet->xdpsocket;
+    message->length = 0;
+
+    ret = xsk_ring_cons__peek(&xdp_socket->rx_ring, 1, &xdp_socket->idx_rx);
+    if (!ret)
+        return UA_STATUSCODE_GOODNODATA; //No packets even after select/poll
+
+    ret = xsk_ring_prod__reserve(&xdp_socket->umem->fq, 1, &xdp_socket->idx_fq);
+    if(ret != 1)
+        return UA_STATUSCODE_BADINTERNALERROR; //Got data but failed to reserve, something's wrong
+
+    addr = xsk_ring_cons__rx_desc(&xdp_socket->rx_ring, xdp_socket->idx_rx)->addr;
+    len  = (UA_UInt32) xsk_ring_cons__rx_desc(&xdp_socket->rx_ring, xdp_socket->idx_rx++)->len;
+    pkt  = (UA_Byte *) xsk_umem__get_data(xdp_socket->umem->buffer, addr);
+
+    /* AF_XDP does not do any filtering on ethertype or protocol.
+     * Manually check for VLAN headers and ETH UADP ethertype.
+     * Note: we use UA_UInt16 to compare ethertype, which is 2-bytes
+     */
+    UA_UInt16 *pkt_proto = (UA_UInt16 *) (pkt + (ETH_ALEN * 2));
+    if(channelDataEthernet->vid > 0 && *pkt_proto == htons(ETHERTYPE_VLAN))
+        pkt_proto += 2;
+
+    if(*pkt_proto != htons(ETHERTYPE_UADP))
+        return UA_STATUSCODE_GOODNODATA;
+
+    eth_hdr = (struct ether_header *) pkt;
+
+    /* Make sure we match our target */
+    if(memcmp(eth_hdr->ether_dhost, channelDataEthernet->targetAddress, ETH_ALEN) != 0) {
+        UA_LOG_ERROR(UA_Log_Stdout, UA_LOGCATEGORY_SERVER,
+                     "Invalid hardware address - XDP socket receive");
+        return UA_STATUSCODE_GOODNODATA;
+    }
+
+    message->data   = (UA_Byte *) pkt + sizeof(struct ether_header);
+    message->length = (size_t) len - sizeof(struct ether_header);
+
+    if(channelDataEthernet->vid > 0) {
+        message->data += 4;
+        message->length -= 4;
+    }
+
+    return UA_STATUSCODE_GOOD;
+}
+#endif
 
 /**
  * Open communication socket based on the connectionConfig.
@@ -175,8 +560,87 @@ UA_PubSubChannelEthernet_open(const UA_PubSubConnectionConfig *connectionConfig)
         return NULL;
     }
 
+    /* get interface index */
+    struct ifreq ifreq;
+    memset(&ifreq, 0, sizeof(struct ifreq));
+    strncpy(ifreq.ifr_name, (char*)address->networkInterface.data,
+            UA_MIN(address->networkInterface.length, sizeof(ifreq.ifr_name)-1));
+
     /* Set default socket options */
-    ethernetSocketOptions sockOptions = {NULL, UA_FALSE, 0, 0};
+    ethernetSocketOptions sockOptions = {NULL, XDP_FLAGS_SKB_MODE, 0, XDP_COPY, UA_FALSE, 0, 0};
+
+    /* XDP parameters - Flags, hw queue option and binding flags */
+    UA_String xdpFlagParam = UA_STRING("xdpflag");
+    UA_String hwReceiveQueueParam = UA_STRING("hwreceivequeue");
+    UA_String xdpBindFlagParam = UA_STRING("xdpbindflag");
+
+    /* Socket priority and socket txtime option */
+    UA_String socketPriority = UA_STRING("sockpriority");
+    UA_String enableSocketTxtime = UA_STRING("enablesotxtime");
+    /* Additional flags for soTxTime */
+    UA_String enableDeadlineMode = UA_STRING("enabledeadlinemode");
+    UA_String enableErrorReport = UA_STRING("enableerrorreport");
+
+    /* iterate over the given KeyValuePair paramters */
+    for(size_t i = 0; i < connectionConfig->connectionPropertiesSize; i++){
+        if(UA_String_equal(&connectionConfig->connectionProperties[i].key.name, &socketPriority)){
+            if(UA_Variant_hasScalarType(&connectionConfig->connectionProperties[i].value, &UA_TYPES[UA_TYPES_UINT32])){
+                sockOptions.socketPriority = (UA_UInt32 *) UA_malloc(sizeof(UA_UInt32));
+                UA_UInt32_copy((UA_UInt32 *) connectionConfig->connectionProperties[i].value.data, sockOptions.socketPriority);
+            }
+        } else if(UA_String_equal(&connectionConfig->connectionProperties[i].key.name, &enableSocketTxtime)){
+            if(UA_Variant_hasScalarType(&connectionConfig->connectionProperties[i].value, &UA_TYPES[UA_TYPES_BOOLEAN])){
+                sockOptions.enableSocketTxTime = *(UA_Boolean *) connectionConfig->connectionProperties[i].value.data;
+            }
+        } else if(UA_String_equal(&connectionConfig->connectionProperties[i].key.name, &enableDeadlineMode)){
+            if(UA_Variant_hasScalarType(&connectionConfig->connectionProperties[i].value, &UA_TYPES[UA_TYPES_BOOLEAN])){
+                if(*(UA_Boolean *) connectionConfig->connectionProperties[i].value.data == UA_TRUE)
+                    sockOptions.sotxtimeDeadlinemode = SOF_TXTIME_DEADLINE_MODE;
+            }
+        } else if(UA_String_equal(&connectionConfig->connectionProperties[i].key.name, &enableErrorReport)){
+            if(UA_Variant_hasScalarType(&connectionConfig->connectionProperties[i].value, &UA_TYPES[UA_TYPES_BOOLEAN])){
+                if(*(UA_Boolean *) connectionConfig->connectionProperties[i].value.data == UA_TRUE)
+                    sockOptions.sotxtimeDeadlinemode = SOF_TXTIME_REPORT_ERRORS;
+            }
+        } else if(UA_String_equal(&connectionConfig->connectionProperties[i].key.name, &xdpFlagParam)){
+            if(UA_Variant_hasScalarType(&connectionConfig->connectionProperties[i].value, &UA_TYPES[UA_TYPES_UINT32])){
+                sockOptions.xdp_flags = *(UA_UInt32 *) connectionConfig->connectionProperties[i].value.data;
+            }
+        } else if(UA_String_equal(&connectionConfig->connectionProperties[i].key.name, &hwReceiveQueueParam)){
+            if(UA_Variant_hasScalarType(&connectionConfig->connectionProperties[i].value, &UA_TYPES[UA_TYPES_UINT32])){
+                sockOptions.hw_receive_queue = *(UA_UInt32 *) connectionConfig->connectionProperties[i].value.data;
+            }
+        } else if(UA_String_equal(&connectionConfig->connectionProperties[i].key.name, &xdpBindFlagParam)){
+            if(UA_Variant_hasScalarType(&connectionConfig->connectionProperties[i].value, &UA_TYPES[UA_TYPES_UINT16])){
+                sockOptions.xdp_bind_flags = *(UA_UInt16 *) connectionConfig->connectionProperties[i].value.data;
+            }
+        } else {
+            UA_LOG_WARNING(UA_Log_Stdout, UA_LOGCATEGORY_SERVER, "PubSub Ethernet Connection creation. Unknown connection parameter.");
+        }
+    }
+
+#if defined(LIBBPF_EBPF)
+    if(connectionConfig->enableXdpSocket) {
+        channelDataEthernet->enableXdpSocket = connectionConfig->enableXdpSocket;
+        channelDataEthernet->xdp_flags = sockOptions.xdp_flags;
+        channelDataEthernet->ifindex = (int)if_nametoindex(ifreq.ifr_name);
+        if((UA_PubSubChannelEthernetXDP_open(channelDataEthernet, &sockOptions, address)) != UA_STATUSCODE_GOOD)
+            return NULL;
+
+        newChannel->sockfd = xsk_socket__fd(channelDataEthernet->xdpsocket->xskfd);
+        newChannel->handle = channelDataEthernet;
+        newChannel->state = UA_PUBSUB_CHANNEL_PUB;
+
+        if(sockOptions.socketPriority)
+            UA_free(sockOptions.socketPriority);
+
+        return newChannel;
+    }
+#endif
+
+    if(connectionConfig->enableXdpSocket)
+        UA_LOG_WARNING(UA_Log_Stdout, UA_LOGCATEGORY_SERVER,
+                       "XDP dependent libraries not found. Use above 5.4 kernel and install libbpf for XDP support. Using AF_PACKET socket");
 
     /* Open a packet socket */
     int sockFd = UA_socket(PF_PACKET, SOCK_RAW, 0);
@@ -244,12 +708,6 @@ UA_PubSubChannelEthernet_open(const UA_PubSubConnectionConfig *connectionConfig)
     }
 #endif
 
-    /* get interface index */
-    struct ifreq ifreq;
-    memset(&ifreq, 0, sizeof(struct ifreq));
-    strncpy(ifreq.ifr_name, (char*)address->networkInterface.data,
-            UA_MIN(address->networkInterface.length, sizeof(ifreq.ifr_name)-1));
-
     if(ioctl(sockFd, SIOCGIFINDEX, &ifreq) < 0) {
         UA_LOG_ERROR(UA_Log_Stdout, UA_LOGCATEGORY_SERVER,
            "PubSub connection creation failed. Cannot get interface index.");
@@ -288,38 +746,6 @@ UA_PubSubChannelEthernet_open(const UA_PubSubConnectionConfig *connectionConfig)
         UA_free(channelDataEthernet);
         UA_free(newChannel);
         return NULL;
-    }
-
-    /* Socket priority and socket txtime option */
-    UA_String socketPriority = UA_STRING("sockpriority");
-    UA_String enableSocketTxtime = UA_STRING("enablesotxtime");
-    /* Additional flags for soTxTime */
-    UA_String enableDeadlineMode = UA_STRING("enabledeadlinemode");
-    UA_String enableErrorReport = UA_STRING("enableerrorreport");
-    /* iterate over the given KeyValuePair paramters */
-    for(size_t i = 0; i < connectionConfig->connectionPropertiesSize; i++){
-        if(UA_String_equal(&connectionConfig->connectionProperties[i].key.name, &socketPriority)){
-            if(UA_Variant_hasScalarType(&connectionConfig->connectionProperties[i].value, &UA_TYPES[UA_TYPES_UINT32])){
-                sockOptions.socketPriority = (UA_UInt32 *) UA_malloc(sizeof(UA_UInt32));
-                UA_UInt32_copy((UA_UInt32 *) connectionConfig->connectionProperties[i].value.data, sockOptions.socketPriority);
-            }
-        } else if(UA_String_equal(&connectionConfig->connectionProperties[i].key.name, &enableSocketTxtime)){
-            if(UA_Variant_hasScalarType(&connectionConfig->connectionProperties[i].value, &UA_TYPES[UA_TYPES_BOOLEAN])){
-                sockOptions.enableSocketTxTime = *(UA_Boolean *) connectionConfig->connectionProperties[i].value.data;
-            }
-        } else if(UA_String_equal(&connectionConfig->connectionProperties[i].key.name, &enableDeadlineMode)){
-            if(UA_Variant_hasScalarType(&connectionConfig->connectionProperties[i].value, &UA_TYPES[UA_TYPES_BOOLEAN])){
-                if(*(UA_Boolean *) connectionConfig->connectionProperties[i].value.data == UA_TRUE)
-                    sockOptions.sotxtimeDeadlinemode = SOF_TXTIME_DEADLINE_MODE;
-            }
-        } else if(UA_String_equal(&connectionConfig->connectionProperties[i].key.name, &enableErrorReport)){
-            if(UA_Variant_hasScalarType(&connectionConfig->connectionProperties[i].value, &UA_TYPES[UA_TYPES_BOOLEAN])){
-                if(*(UA_Boolean *) connectionConfig->connectionProperties[i].value.data == UA_TRUE)
-                    sockOptions.sotxtimeDeadlinemode = SOF_TXTIME_REPORT_ERRORS;
-            }
-        } else {
-            UA_LOG_WARNING(UA_Log_Stdout, UA_LOGCATEGORY_SERVER, "PubSub Ethernet Connection creation. Unknown connection parameter.");
-        }
     }
 
     /* Setting the socket priority to the socket */
@@ -394,6 +820,9 @@ UA_PubSubChannelEthernet_regist(UA_PubSubChannel *channel,
     UA_PubSubChannelDataEthernet *channelDataEthernet =
         (UA_PubSubChannelDataEthernet *) channel->handle;
 
+    if(channelDataEthernet->enableXdpSocket)
+        return UA_STATUSCODE_GOOD; // Regist not required for XDP sockets
+
     if(!is_multicast_address(channelDataEthernet->targetAddress))
         return UA_STATUSCODE_GOOD;
 
@@ -421,6 +850,9 @@ UA_PubSubChannelEthernet_unregist(UA_PubSubChannel *channel,
                                   UA_ExtensionObject *transportSettings) {
     UA_PubSubChannelDataEthernet *channelDataEthernet =
         (UA_PubSubChannelDataEthernet *) channel->handle;
+
+    if(channelDataEthernet->enableXdpSocket)
+        return UA_STATUSCODE_GOOD; // Unregist not required for XDP sockets
 
     if(!is_multicast_address(channelDataEthernet->targetAddress)) {
         return UA_STATUSCODE_GOOD;
@@ -516,6 +948,12 @@ UA_PubSubChannelEthernet_send(UA_PubSubChannel *channel,
     UA_PubSubChannelDataEthernet *channelDataEthernet =
         (UA_PubSubChannelDataEthernet *) channel->handle;
 
+#if defined LIBBPF_EBPF
+    if(channelDataEthernet->enableXdpSocket) {
+        UA_StatusCode retval = UA_PubSubChannelEthernetXDP_send(channelDataEthernet, buf);
+        return retval;
+    }
+#endif
     /* Allocate a buffer for the ethernet data which contains the ethernet
      * header (without VLAN tag), the VLAN tag and the OPC-UA/Ethernet data. */
     char *bufSend, *ptrCur;
@@ -625,6 +1063,14 @@ UA_PubSubChannelEthernet_receive(UA_PubSubChannel *channel, UA_ByteString *messa
             return UA_STATUSCODE_BADINTERNALERROR;
         }
     }
+
+#if defined LIBBPF_EBPF
+    if(channelDataEthernet->enableXdpSocket) {
+        retval = UA_PubSubChannelEthernetXDP_receive(channelDataEthernet, message);
+        return retval;
+    }
+#endif
+
 #ifdef UA_ARCHITECTURE_VXWORKS
     clock_gettime(CLOCK_REALTIME, &currentTime);
 #else
@@ -726,6 +1172,18 @@ UA_PubSubChannelEthernet_receive(UA_PubSubChannel *channel, UA_ByteString *messa
     return retval;
 }
 
+static UA_StatusCode
+UA_PubSubChannelEthernet_release(UA_PubSubChannel *channel) {
+#if defined(LIBBPF_EBPF)
+    UA_PubSubChannelDataEthernet *channelDataEthernet = (UA_PubSubChannelDataEthernet *) channel->handle;
+    xdpsock *xdp_socket = channelDataEthernet->xdpsocket;
+    xsk_ring_prod__submit(&xdp_socket->umem->fq, 1);
+    xsk_ring_cons__release(&xdp_socket->rx_ring, 1);
+    xdp_socket->rx_npkts += 1;
+#endif
+    return UA_STATUSCODE_GOOD;
+}
+
 /**
  * Close channel and free the channel data.
  *
@@ -733,8 +1191,23 @@ UA_PubSubChannelEthernet_receive(UA_PubSubChannel *channel, UA_ByteString *messa
  */
 static UA_StatusCode
 UA_PubSubChannelEthernet_close(UA_PubSubChannel *channel) {
+    UA_PubSubChannelDataEthernet *channelDataEthernet = (UA_PubSubChannelDataEthernet *) channel->handle;
+#if defined(LIBBPF_EBPF)
+    if(channelDataEthernet->enableXdpSocket) {
+        xsk_socket__delete(channelDataEthernet->xdpsocket->xskfd);
+        (void)xsk_umem__delete(channelDataEthernet->xdpsocket->umem->umem);
+        /* Detach XDP program from the interface */
+        bpf_set_link_xdp_fd(channelDataEthernet->ifindex, -1, channelDataEthernet->xdp_flags);
+        UA_free(channelDataEthernet->xdpsocket->umem);
+        UA_free(channelDataEthernet->xdpsocket);
+        UA_free(channelDataEthernet);
+        UA_free(channel);
+        return UA_STATUSCODE_GOOD;
+    }
+#endif
+
     UA_close(channel->sockfd);
-    UA_free(channel->handle);
+    UA_free(channelDataEthernet);
     UA_free(channel);
     return UA_STATUSCODE_GOOD;
 }
@@ -755,6 +1228,8 @@ TransportLayerEthernet_addChannel(UA_PubSubConnectionConfig *connectionConfig) {
         pubSubChannel->send = UA_PubSubChannelEthernet_send;
         pubSubChannel->receive = UA_PubSubChannelEthernet_receive;
         pubSubChannel->close = UA_PubSubChannelEthernet_close;
+        if(connectionConfig->enableXdpSocket)
+            pubSubChannel->release = UA_PubSubChannelEthernet_release;
         pubSubChannel->connectionConfig = connectionConfig;
     }
     return pubSubChannel;
