+++ src/pubsub/ua_pubsub_reader.c
@@ -946,15 +946,12 @@ void UA_ReaderGroup_subscribeCallback(UA_Server *server, UA_ReaderGroup *readerG
 
                 previousPosition = currentPosition;
             } while((buffer.length) > currentPosition);
-#ifdef UA_ENABLE_PUBSUB_ETH_UADP_XDP
-            if (connection->channel->release)
-                connection->channel->release(connection->channel);
 
-            if (!connection->channel->release)
+            if(connection->channel->release)
+                connection->channel->release(connection->channel);
+            else
                 UA_ByteString_clear(&buffer);
-#else
-            UA_ByteString_clear(&buffer);
-#endif
+
             return;
 
         } else {
@@ -979,18 +976,13 @@ void UA_ReaderGroup_subscribeCallback(UA_Server *server, UA_ReaderGroup *readerG
                 previousPosition = currentPosition;
             } while((buffer.length) > currentPosition);
         }
-#ifdef UA_ENABLE_PUBSUB_ETH_UADP_XDP
-        if (connection->channel->release)
-            connection->channel->release(connection->channel);
-#endif
-    } 
 
-#ifdef UA_ENABLE_PUBSUB_ETH_UADP_XDP
-    if (!connection->channel->release)
-        UA_ByteString_clear(&buffer);	
-#else
-    UA_ByteString_clear(&buffer);
-#endif
+        if(connection->channel->release)
+            connection->channel->release(connection->channel);
+    }
+
+    if(!connection->channel->release)
+        UA_ByteString_clear(&buffer);
 }
 
 /* Add new subscribeCallback. The first execution is triggered directly after
